Redid and cleaned app structure with major overhaul + Setup button increment project with test with notes about structure of development of both the main app and unit testing
